refactored jobs and job cards and moved mapping to parent page in jobs
